requirements.txt
langchain==1.2.10
langchain-google-genai==4.2.1
python-dotenv==1.2.1
requests==2.32.5